Moves track metric telemetry hooks etc to shared observability plugin
