+++ x-pack/plugins/infra/public/index.ts
@@ -16,12 +16,5 @@ export const plugin: PluginInitializer<
   return new Plugin(context);
 };
 
-export {
-  useTrackPageview,
-  useUiTracker,
-  UiTracker,
-  TrackMetricOptions,
-  METRIC_TYPE,
-} from './hooks/use_track_metric';
 export { FORMATTERS } from './utils/formatters';
 export { InfraFormatterType } from './lib/lib';
