+++ x-pack/plugins/observability/public/hooks/use_track_metric.tsx
