+++ x-pack/plugins/observability/public/index.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export {
+  useTrackPageview,
+  useUiTracker,
+  UiTracker,
+  TrackMetricOptions,
+  METRIC_TYPE,
+} from './hooks/use_track_metric';
